sched/wait: Add swq_has_sleeper()
Which is the equivalent of what we have in regular waitqueues.
I'm not crazy about the name, but this also helps us get both
apis closer -- which iirc comes originally from the -net folks.

We also duplicate the comments for the lockless swait_active(),
from wait.h. Future users will make use of this interface.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Dede Dindin Qudsy <xtrymind@gmail.com>
Signed-off-by: fajar4561 <maulanafajar751@gmail.com>